+++ packages/api-explorer/src/components/ExploreType/ExploreProperty.spec.tsx
@@ -29,8 +29,9 @@ import { fireEvent, screen, waitFor } from '@testing-library/react'
 import { IdeFileManifest } from '@looker/icons'
 import { Toc } from '@styled-icons/material/Toc'
 import { Tag } from '@styled-icons/material-rounded/Tag'
+
 import { api } from '../../test-data'
-import { renderWithSearchAndRouter } from '../../test-utils'
+import { renderWithRouter } from '../../test-utils'
 import { ExploreProperty, ExplorePropertyDetail, typeIcon } from '.'
 
 describe('ExploreProperty', () => {
@@ -41,7 +42,7 @@ describe('ExploreProperty', () => {
       test('Shows read-only property and description', async () => {
         const property = type.properties.id
         expect(property).toBeDefined()
-        renderWithSearchAndRouter(
+        renderWithRouter(
           <ExplorePropertyDetail api={api} property={property} />
         )
         expect(property.deprecated).toEqual(false)
@@ -61,7 +62,7 @@ describe('ExploreProperty', () => {
       test('Shows required property and description', async () => {
         const property = type.properties.dashboard_id
         expect(property).toBeDefined()
-        renderWithSearchAndRouter(
+        renderWithRouter(
           <ExplorePropertyDetail api={api} property={property} />
         )
         expect(property.deprecated).toEqual(false)
@@ -84,9 +85,7 @@ describe('ExploreProperty', () => {
     test('int64 property icon is correct', () => {
       const property = type.properties.row
       expect(property.type.jsonName).toEqual('int64')
-      renderWithSearchAndRouter(
-        <ExploreProperty api={api} property={property} />
-      )
+      renderWithRouter(<ExploreProperty api={api} property={property} />)
       expect(screen.getByTitle('int64')).toBeInTheDocument()
       const legend = typeIcon(property.type)
       expect(legend).toEqual({ icon: <Tag />, title: 'int64' })
@@ -95,9 +94,7 @@ describe('ExploreProperty', () => {
     test('array property icon is correct', () => {
       const property = type.properties.listens_to_filters
       expect(property.type.jsonName).toEqual('string[]')
-      renderWithSearchAndRouter(
-        <ExploreProperty api={api} property={property} />
-      )
+      renderWithRouter(<ExploreProperty api={api} property={property} />)
       expect(screen.getByTitle(property.type.jsonName)).toBeInTheDocument()
       const legend = typeIcon(property.type)
       expect(legend).toEqual({ icon: <Toc />, title: 'string[]' })
@@ -106,9 +103,7 @@ describe('ExploreProperty', () => {
     test('hash property icon is correct', () => {
       const property = type.properties.field
       expect(property.type.jsonName).toEqual('Hash[any]')
-      renderWithSearchAndRouter(
-        <ExploreProperty api={api} property={property} />
-      )
+      renderWithRouter(<ExploreProperty api={api} property={property} />)
       expect(screen.getByTitle(property.type.jsonName)).toBeInTheDocument()
       const legend = typeIcon(property.type)
       expect(legend).toEqual({ icon: <IdeFileManifest />, title: 'Hash[any]' })