+++ packages/api-explorer/src/reducers/index.ts
@@ -24,4 +24,3 @@
 
  */
 export * from './spec'
-export * from './search'