+++ packages/hackathon/src/scenes/JudgingEditorScene/index.ts
@@ -23,4 +23,4 @@
  SOFTWARE.
 
  */
-export { JudgingEditorScene } from './JudgingEditorScene'
+export * from './JudgingEditorScene'