+++ packages/hackathon/src/scenes/ProjectViewScene/index.ts
@@ -23,31 +23,4 @@
  SOFTWARE.
 
  */
-
-import { useState, useEffect } from 'react'
-import type { TagList, TypeList } from '@looker/sdk-codegen'
-
-export const useDebounce = (value: any, delay: number) => {
-  const [debouncedValue, setDebouncedValue] = useState()
-  useEffect(() => {
-    // Update debounced value only once delay has been elapsed
-    const handler = setTimeout(() => {
-      setDebouncedValue(value)
-    }, delay)
-    // Reset the timeout on receiving a new keyword
-    return () => {
-      clearTimeout(handler)
-    }
-  }, [value])
-  return debouncedValue
-}
-
-export const countMethods = (tags: TagList) => {
-  let result = 0
-  Object.values(tags).forEach((methods) => {
-    result += Object.entries(methods).length
-  })
-  return result
-}
-
-export const countTypes = (types: TypeList) => Object.entries(types).length
+export * from './ProjectViewScene'