+++ packages/run-it/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+### [0.9.22](https://www.github.com/looker-open-source/sdk-codegen/compare/run-it-v0.9.21...run-it-v0.9.22) (2021-10-27)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @looker/sdk bumped from ^21.18.0 to ^21.18.1
+    * @looker/sdk-codegen bumped from ^21.1.0 to ^21.1.1
+    * @looker/code-editor bumped from ^0.1.12 to ^0.1.13
+    * @looker/sdk-rtl bumped from ^21.1.0 to ^21.1.1
+
 ### [0.9.21](https://www.github.com/looker-open-source/sdk-codegen/compare/run-it-v0.9.20...run-it-v0.9.21) (2021-10-24)